Fix and organize type definition issues
