--- src/core/assets/asset-handler/meta-schemas/userDatas.d.ts
@@ -1,27 +1,0 @@
-// 这个文件用于记录导入器的各种类型定义，导出声明文件的时候将导出这个配置文件
-export type ImageImportType = 'raw' | 'texture' | 'normal map' | 'sprite-frame' | 'texture cube';
-
-/** 图片资源的 userData  */
-export interface ImageAssetUserData {
-    /** 图片类型 */
-    type: ImageImportType;
-    /** 垂直翻转 */
-    flipVertical?: boolean;
-    /** 消除透明伪影 */
-    fixAlphaTransparencyArtifacts?: boolean;
-    /** 是否为 RGBE */
-    isRGBE?: boolean;
-    /** 这个图片是不是拥有 alpha 通道 */
-    hasAlpha?: boolean;
-    /** 重定向的 uuid */
-    redirect?: string;
-    visible?: boolean;
-    /** 是否翻转绿通道 */
-    flipGreenChannel?: boolean;
-
-    /**
-     * 部分资源导入后可能产生多张图像资源
-     */
-    sign?: string;
-    alpha?: string;
-}
